turn::relay::relay_range

Struct RelayAddressGeneratorRanges

pub struct RelayAddressGeneratorRanges {
    pub relay_address: IpAddr,
    pub min_port: u16,
    pub max_port: u16,
    pub max_retries: u16,
    pub address: String,
    pub net: Arc<Net>,
}

RelayAddressGeneratorRanges can be used to only allocate connections inside a defined port range.

Fields
relay_address: IpAddr

relay_address is the IP returned to the user when the relay is created.

min_port: u16

min_port the minimum port to allocate.

max_port: u16

max_port the maximum (inclusive) port to allocate.

max_retries: u16

max_retries the amount of tries to allocate a random port in the defined range.

address: String

address is passed to Listen/ListenPacket when creating the Relay.

net: Arc<Net>